change user repo class

+++ src/main/java/com/java/crv/BibleReaderCommentary/repositories/UserRepository.java
@@ -1,9 +1,8 @@
 package com.java.crv.BibleReaderCommentary.repositories;
 
 import org.springframework.data.repository.CrudRepository;
+import com.java.crv.BibleReaderCommentary.domain.*;
 
-import com.java.crv.BibleReaderCommentary.domain.User;
-
-public interface UserRepository extends CrudRepository<User, Integer>{
+public interface UserRepository extends CrudRepository<User, Long>{
 
 }
